project/pi/requirements.txt
flask==2.1.1
requests==2.27.1
paho-mqtt==1.6.1
schedule==1.1.0
smbus2
RPI.GPIO
